openai/requirements.txt
autopep8==1.6.0
backoff==2.2.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Deprecated==1.2.13
et-xmlfile==1.1.0
Flask==2.0.2
googleapis-common-protos==1.58.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.21.3
openai==0.19.0
openpyxl==3.0.9
opentelemetry-api==1.15.0
opentelemetry-exporter-otlp-proto-http==1.15.0
opentelemetry-instrumentation==0.36b0
opentelemetry-instrumentation-flask==0.36b0
opentelemetry-instrumentation-requests==0.36b0
opentelemetry-instrumentation-wsgi==0.36b0
opentelemetry-proto==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
opentelemetry-util-http==0.36b0
pandas==1.3.4
pandas-stubs==1.2.0.35
protobuf==4.21.12
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six==1.16.0
toml==0.10.2
tqdm==4.62.3
types-pytz==2022.6.0.1
typing-extensions==4.4.0
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.14.1
